Separated Menu Group into two classes - one with default menu that is activated whenever another menu desactivates, and other without default menu
