+++ Project Anolis/Assets/Scripts/Logistics/MenuGroups.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b9382c43460482d4fb558ff299014a9c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
