+++ Project Anolis/Assets/Scripts/Logistics/MenuGroups/MenuGroupWithDefault.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c36e9a47d38f98e4dbbba4711b2f7be5
+guid: d4cefbd03ab93ca4b9d317d2ad948756
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
