--- Project Anolis/Assets/Scripts/Logistics/Menus/Implementations/EmptyMenu.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace Logistics
-{
-    public class EmptyMenu : Menu
-    {
-        private void Awake()
-        {
-            Ui = new GameObject();
-        }
-
-        public override bool CanHandleSelection()
-        {
-            return false;
-        }
-    }
-}
